apps/notification_app/tests/factories/push_notification_request_factory.py
import factory

from ...models import PushNotificationRequest


class PushSubscriptionRequestFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = PushNotificationRequest
        